use long math for long values

+++ src/java/org/apache/cassandra/utils/SlabAllocator.java
@@ -115,7 +115,7 @@ public class SlabAllocator extends Allocator
      */
     public long getMinimumSize()
     {
-        return unslabbed.get() + (regionCount - 1) * REGION_SIZE;
+        return unslabbed.get() + (regionCount - 1) * (long)REGION_SIZE;
     }
 
     /**
@@ -123,7 +123,7 @@ public class SlabAllocator extends Allocator
      */
     public long getMaximumSize()
     {
-        return unslabbed.get() + regionCount * REGION_SIZE;
+        return unslabbed.get() + regionCount * (long)REGION_SIZE;
     }
 
     /**
